Move bannerDate under banner folder
