+++ src/components/banner/bannerDate/_spec.js
