+++ src/components/banner/bannerDate/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
-import { media } from '../../theme/media'
-import theme from '../../theme/theme'
+import { media } from '../../../theme/media'
+import theme from '../../../theme/theme'
 
 export const BannerDate = styled.p`
   color: #ffffff;
